[graphite] Optimize drawDRRect and nested rect contours
Detects nested rect contours in drawPath and sends them to drawDRRect.
DrawDRRect has 3 optimized paths to avoid path rendering that don't
significantly change the pipeline that would be required:

1. If the inner rect fully cuts out a side of the outer rect, it becomes
a drawRect(). 2. If the outer - inner rrect is a stroke, it becomes a
stroked [r]rect. 3. Otherwise, it's a drawRRect(outer) inside a
clipRRect(inner).

These all end up mapping to the AnalyticRRectRenderStep and/or
CoverBoundsRenderStep, just like a regular drawRRect/drawRect would. The
minor exception is #3 that can add an AnalyticClip option.

These options are skipped if there's a different fill style or a path
effect.

--- This pattern comes up in GMail's captured SKP. With graphite before
this CL, we triggered MSAA and had 206 draw calls and frame time of
402us on my M3.

With this change, we no longer trigger MSAA, it goes down to 111 draw
calls and a frame time of 328us. ---

I considered having this logic exist inside the ClipStack to try and
apply difference clips to draws, or intersect clips to inverse-filled
shapes that could produce stroke geometry. However, it quickly became
cumbersome and complicated, and in the non-drawDRRect scenarios I
thought I saw the pattern, the clip shapes were actually already reduced
and the analytic geometry wasn't a stroke (it just looked like a stroke
after the scissor test).

Change-Id: I30198ae6dda9448c1b41f9ab7b3076f5427cb652
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1025999
Reviewed-by: Nicolette Prevost <nicolettep@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>