[graphite] Preserve rounded corners for stroke bounds in ClipStack
At least in some webpage patterns from Blink, there is a clipRRect that
matches the outer edge of a stroked circular border. Without preserving
the roundedness, we'd end up thinking the outset bounds of the stroked
circle would be clipped and add an AnalyticClip.

Change-Id: I4ac731be54af052fe8c34070d23abfea25c4b0a5
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1027316
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>