[graphite] Add option to use raster atlas for small paths
A number of website SKPs feature small icons that trigger MSAA
path rendering, but cover a fraction of the entire render target.
Even with memoryless MSAA attachments, there is a performance hit
that impacts the rest of the render pass.

While this is mitigated by tiling, the GPU tiles in chrome are still
quite large. This adds an option that lets us fall back to the raster
path atlas for paths smaller than a certain size. While by default
this won't impact renderer selection, I plan to run a finch experiment
to see if it produces a net positive.

Change-Id: Id1e3bc70da866a0a33dc68a07a61c967703603d1
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1028721
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>