Remove alignment constraint on SkMemoryCtxPatch
The compiler was generating some code in SkRasterPipeline::run
like:
    vmovaps %ymm0,0xc0(%rcx,%rdi,1)
which assumes memory that is aligned to 32 bytes (256 bits) but
this was not the case because AutoSTMalloc allocates an internal
buffer that is only 4 byte aligned (it's a buffer of uint32_t).

By removing this, the compiler has to generate the unaligned version
    vmovups %ymm0,0xe0(%rcx,%rdi,1)
which may be slower, but doesn't crash.

Change-Id: Ie61ffe7d2b14b0319cb2e15693243a4db10f30f0
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1030296
Reviewed-by: Ben Wagner <bungeman@google.com>
Commit-Queue: Kaylee Lubick <kjlubick@google.com>