Set up scripts for louhi flows for binary-size and cmake-release
skia-with-swift-shader-base is not used anymore, but binary-size
and cmake-release is. This moves said Dockerfiles and adds build.sh
scripts so that we can add Louhi flows to build them again when
we need to update them.

Change-Id: I9d2a0656d9bca9fb64c12e507abeaccd30c48611
No-Try: true
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1029945
Auto-Submit: Kaylee Lubick <kjlubick@google.com>
Reviewed-by: Eric Boren <borenet@google.com>
Commit-Queue: Eric Boren <borenet@google.com>