[graphite] Ensure scissor has integer values
The fScissor value is stored as a Rect for convenience when processing
the draw and clip elements but it must hold integer values. This was
guaranteed for the initial scissor that came from `snap_scissor`.
However, when applying a tight scissor for a given clip element,
`applyScissor()` was called with the clip's original rect coords.

At this point, we knew e.shape().rect().nearlyEquals(rounded) so they
were integers up to an epsilon.  You can construct a nearly 1px wide
clip rect like (x+eps, top, x + 1 - eps, bot) that is not empty and
won't trigger an early clipped-out event. Indeed it should draw something
if the draw overlaps that 1px opening.

If eps is small enough, it'll pass nearlyEqual() checks vs (x, top, x+1, bot)
assuming x, top, and bot are integers. The DrawShape::fScissor at this
point is non-empty and will leave the Clip::fDrawBounds value as non-empty
so it will return false from Clip::isClippedOut().  However, the scissor
is cast to integers for storing as an SkIRect and with this construction,
(x+eps, top, x + 1 - eps, bot) becomes (x, top, x, bot) and is now
incorrectly seen as empty.

Since the drawBounds weren't empty, DrawContext asserts when we try to
record the draw. On a release build, we'll do extra work and then end
up setting an empty scissor that prevents it from being rendered
(although it should render a thin strip).

This CL adds additional asserts closer to the scissor application and
to the Clip object creation, and changes it so that the tight scissor
handling passes in the rounded value so that when the shape is nearly
equal to integers, we use the actual integer values for the scissor.

Bug: b/436691546
Change-Id: I5d24f2dca678453e873275b7c53db3a22c8830ec
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1033683
Auto-Submit: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Thomas Smith <thomsmit@google.com>
Reviewed-by: Thomas Smith <thomsmit@google.com>