[M139] [codecs] Avoid unneeded color space transforms
For certain color types (f16, 101010) SkCodec::initializeColorXform sets
up an skcms transform even in the absence of a destination color space,
in order to handle pixel format conversions.

When a dst color space is not specified, we reuse the source icc profile
as dst to avoid unnecessary color space conversions.

The problem is we make a copy of the icc profile, while skcms_Transform
only performs a shallow (pointer) profile comparison [1] when deciding
whether color translation is needed.

This has a couple of issues:

1) we are unnecessarily round-tripping colors

2) skcms_Transform only supports certain dst profile types (notably
   it does not support icc profiles with table-based TRCs [2]), and
   since the src profile is controlled by the image asset, we are
   vulnerable to inputs that can cause skcms_Transform to fail
   (and to trigger asserts in debug builds).

The proposed fix ensures that when we're attempting to skip color
translation by passing the same src and dst profile, we pass the same
pointer for both. This allow skcms_Transform to skip color transform and
avoids the issues mentioned above.

[1]
https://source.chromium.org/chromium/chromium/src/+/main:third_party/skia/modules/skcms/skcms.cc;drc=aae31cfd0a4d152bacdea12c9eb5c4416d7c79a1;l=2752
[2]
https://source.chromium.org/chromium/chromium/src/+/main:third_party/skia/modules/skcms/skcms.cc;drc=aae31cfd0a4d152bacdea12c9eb5c4416d7c79a1;l=2614

Bug: 436624847, 437355632
Change-Id: I182125563e0a8bc74b33eda92113a2388e3c0dcb
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1034237
Reviewed-by: Daniel Dilan <danieldilan@google.com>
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1034696
Commit-Queue: Florin Malita <fmalita@google.com>