[pdf] Unpremul after interpolating unpremul
In "[pdf] Respect kInterpolateColorsInPremul_Flag" the color values were
properly premultiplied before interpolation. However, what this change
missed is that the outputs (to PDF) need to be in unpremultiplied. This
means passing along and interpolating the alpha with the color channels
so that they can be unpremultiplied.

[0] https://skia.googlesource.com/skia.git/+/bead0db0bd33fb51b7ca7f2809e56be36a565ade

Bug: 437334351
Change-Id: Ia37beb9b01729c5ffd3255bdaa06a3f1c1488e0e
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1036037
Reviewed-by: Florin Malita <fmalita@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>