Make skslc use Tint built from CMake instead of GN rules
As a first step in getting Skia to use CMake for building
all of Dawn, I started first with just Tint. Earlier attempts
built Tint's ~90 lib_tint_foo.a files and listed them all in
third_party/dawn/BUILD.gn but that was brittle and didn't link
correctly with normal `ld` (it would have required extra precise
ordering of them or using `lld`). There was no way I could figure
out to use Dawn/Tint's existing rules to create a single static Tint
archive, so I settled on using ar to make one.

The trickiest logic for this CL is using ninja to generate a .d
dependency file for consumption by GN to have smart incremental
builds. Neither CMake nor ninja have built in ways of doing this,
but by using ninja -tinputs to get a list of all targets needed
to build the targets we want and then ninja -tdeps to get all
the dependent files for said targets, we can create one and
pipe it to GN as tint_combined.a.d.

This uses a similar approach to bazel_build.py [1] to check the
output library's hash before copying it to avoid confusing GN
with the updated timestamps.

[1] https://github.com/google/skia/blob/47f8739423876471d8bf618e6daf537cd0e7d67c/gn/bazel_build.py

Bug: b/434977307
Change-Id: Ia0c36a670bea841e3671bdf8b978cbee4e5513b7
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1031656
Reviewed-by: Ben Wagner <bungeman@google.com>
Reviewed-by: Eric Boren <borenet@google.com>
Commit-Queue: Kaylee Lubick <kjlubick@google.com>