Fix Android NDK FontMgr alias scanning
The previous approach of comparing what was read from the fonts.xml and
what the AFont reported works through Android 14. Unfortunately,
starting in Android 15 the behavior of AFont changed to not report a
variation position if it was the default position for an axis.

Make this more robust by creating the underlying typeface from the AFont
information and (after matching on data) a clone typeface from the xml
information. Then check if these report the same variation position.

Change-Id: Ie5301b7f98b8472b719cb2a2e459c3c2a5e77ada
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1037598
Commit-Queue: Ben Wagner <bungeman@google.com>
Reviewed-by: Julia Lavrova <jlavrova@google.com>
Reviewed-by: Florin Malita <fmalita@google.com>