Update SkFont size documentation
The SkFont documentation currently claims that the size is in "points",
which is confusing since it isn't 1/72 of anything. Instead the size on
an SkFont is in the same units as any value passed to an SkCanvas, which
are documented as "local coordinate units". The value describes the size
to which the EM unit in the font must be scaled.

Bug: 427932827
Change-Id: Iee43d3a0cdbc6b1191f50c93b80966b1a94d8120
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1038456
Commit-Queue: Ben Wagner <bungeman@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>