[graphite] PatchWriter handles caps directly
PatchWriter is updated to be able to write squares, encoded similarly
to how circle caps are encoded. These squares are either standalone
squares, or oriented half squares that join with a prior patch.
This allows deferring all calculations for the square vertices to the
GPU, so works with hairlines or regular strokes.

The deferred patch writing system in PatchWriter is adjusted to keep
track of the first and last control point, and whether or not it had
verbs recorded (even if they were degenerate). Writing the deferred
patch then can handle either joining or adding caps.

This allows the TessellatedStrokeRenderStep to iterate the path data
directly instead of sending everything through the StrokeIter class,
which previously added extra verbs and had a queue of control points
to be able to handle joins and caps. However, it would push and
pop all verbs and path data for an SkPath leading to performance
hot spots. With the caps handled by just the deferred patch in
PatchWriter, it only adds extra work when actually ending a
contour.

On my M3 air, motionmarkpaths.skp goes from 2.83ms to 2.35ms (17%).

Lastly, this also sets the stage for two improvements:
1. Non path Shapes can more easily be stroked using the raw path
API once that's stable.
2. Perspective will hopefully be easier to implement since caps
no longer depend on mapping control points on the CPU.

Bug: b/238768754
Change-Id: I7e384f4f61e77d5a8644d268e936dda9b4776d04
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1031736
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>