Add default case for color encode
Now that we can technically accept any SkColorType for SkPngEncoder,
this enumeration may be wrong. Add a fail in debug in case one is to add
a color type, this can alert them to add the enumeration here. Otherwise
if we haven't already enumerated it, don't set incorrect sigBits.

Change-Id: Iad0ef6168083de86e79edd8f955d5292ab2b9de0
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1038856
Reviewed-by: Florin Malita <fmalita@google.com>
Commit-Queue: Daniel Dilan <danieldilan@google.com>