Add output colorspace parameter to SkShader::makeWithWorkingColorSpace
Internally, this updates the SkWorkingColorSpaceShader to hold
state for working in unpremul values, and adjusts the backend handling
to use this information.

However, since working in unpremul will also require all inputs to
convert or produce unpremul values, this aspect of the feature does
not work completely correctly. It is not exposed in the public API
yet (just the input/output CS parameters), but the GM contains cases
for working in unpremul to show the current failures.

Bug: b/431253869
Change-Id: Ie45286389c033249ce84a1746dae5f7bed0c8d76
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1021417
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Florin Malita <fmalita@google.com>