[graphite][vulkan] Dedicated uint32 for usage flags in texture key
If a texture is imported by Skia, any bit in VkImageUsageFlags could be
set. A whole uint32 is dedicated in this change to the usage flags when
creating the texture key.

This is in preparation for a change that uses VK_EXT_host_image_copy,
setting a high bit in this uint32.

Bug: skia:435241465
Change-Id: I4f719d571f382765b0e157f6442458bcb4ca9b59
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1040837
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Nicolette Prevost <nicolettep@google.com>
Commit-Queue: Shahbaz Youssefi <syoussefi@google.com>