Properly handle value types in LRUCache
Provide a way to correctly forward values into the creation of Entry.
This allows for move only keys and values and also avoids any extra
copies for other types. This also corrects Entry::fKey to be const, as
the key must not be changed.

Change-Id: I11800c6c2216984b20e3a4ed0eb90674031466cc
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1042537
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>