[graphite][vulkan] Remove VK_ACCESS_HOST_WRITE_BIT from src flags
The Host Write Ordering Guarantees section of the spec [1] specifies
that vkQueueSubmit implicitly includes a _visibility_ operation (cache
invalidate) between host writes and all device access. This means that
VK_ACCESS_HOST_WRITE_BIT is unnecessary.

[1]:https://docs.vulkan.org/spec/latest/chapters/synchronization.html#synchronization-submission-host-writes

Note that the bit is necessary if host write happens after queue
submission and synchronization is done with e.g. a host-accessible
VkEvent, i.e. it's a pretty rare situation.

Note also that the _availability_ operation (cache flush) is not
implicit, which is what vkFlushMappedMemoryRanges is for.

Bug: skia:441128467
Change-Id: I53ebf581597f50bb264560dd19c88ab1003a9afe
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1045776
Reviewed-by: Nicolette Prevost <nicolettep@google.com>
Commit-Queue: Shahbaz Youssefi <syoussefi@google.com>