Hack RobotoStatic in SkFontMgr_NDK
Some Android versions have a variable Roboto font named Roboto but also
use a font named RobotoStatic (which does not claim to be Roboto) for
the 400 weight.  If RobotoStatic is found but does not have the name
"Roboto", add it.  Fixed in Android U "[2nd attempt] Revive use of VF
font for regular style of roboto font" [0]

[0] https://android.googlesource.com/platform/frameworks/base/+/89abe560d722a6f4136b7a05d80f23b269413aad

Bug: 40618213
Change-Id: I2f5d9d8a5b71b7d7f00896b88f37968d4ef6c2d3
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1047456
Reviewed-by: Julia Lavrova <jlavrova@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>