[rustpng] Use newer rust api `read_row` to avoid extra memcpys
Uses image-rs png `read_row` [1], which Lukasz added for this purpose.
On the CPP side, two new functions are added. canReadRow() checks if we
can decode straight from image-rs to dst, and if we can we use
incrementalDecode(), the original was renamed to
incrementalDecodeColorXForm(). This is essentially the same function,
without any memcpys. An extra parameter is added to
expandDecodedInterlacedRow() to check if we can skip another memcpy. I
thought it was most readable to keep these as two different functions.
Locally I see slight speed ups but curious to see perf.

[1]
https://docs.rs/png/0.18.0-rc.3/png/struct.Reader.html#method.read_row

Bug: 399891492
Change-Id: I20812f959069009c38d8758b19ed7f1100f0407c
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1044996
Reviewed-by: Kaylee Lubick <kjlubick@google.com>
Reviewed-by: Florin Malita <fmalita@google.com>
Commit-Queue: Daniel Dilan <danieldilan@google.com>