SkAndroidCodec: Remove one-off CICP implementation
SkAndroidCodec added support for CICP before CICP parsing was added to
SkColorSpace. Remove it and point it at the new version.

The behavior of PQ color spaces does not change.

The behavior of HLG color spaces does change. It is still the case that
75% signal maps to 1. But now, 100% signal maps to 1000/203, and the
reference OOTF is respected.

This causes some test failures, because the tests expected a precise
toXYZD50 matrix. To that end, add the matrix (via pointer) to
SkNamedPrimaries::TableEntry, to avoid the extra computation of the
matrix from the primaries. It may be worth doing this for all entries,
to avoid the extra computation.

Bug: skia:420956739
Change-Id: I3fde9af9a4d75148d8aa94db8741649c4bafb768
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1048476
Reviewed-by: Kaylee Lubick <kjlubick@google.com>
Commit-Queue: Christopher Cameron <ccameron@google.com>
Reviewed-by: Christopher Cameron <ccameron@google.com>