Add script for automatically building emsdk-base docker container
I'd like to update this container for both better WebGPU support
and to include ninja (which should make the linked bug easier).

I'd also like to combine the emsdk-base and canvaskit-emsdk images
together, there doesn't really need to be the second one (which is
why libfreetype is being included here)

This CL (https://louhi-config-internal-review.git.corp.google.com/c/skia-infra/+/233740)
adds a Louhi flow to build this image in a hermetic way and then
a second CL will start using that image.

Change-Id: I4d4c1e631f68eee3d5d217f09a452bb99ad2b403
Bug: b/434977307
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1048636
Reviewed-by: Eric Boren <borenet@google.com>