Better bounds check idIndex in ImGuiLayer
If the paint pointer references the upper half of memory the intptr_t
will be negative, which is not a value that is intended as an index into
fSkiaWidgetFuncs. Fixing this bounds check avoids a crash in viewer on
Android.

Change-Id: I0f90cca53d8ea01333bf1bedf5464aa2e074b9b4
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1051338
Commit-Queue: Ben Wagner <bungeman@google.com>
Reviewed-by: Florin Malita <fmalita@google.com>
Commit-Queue: Florin Malita <fmalita@google.com>
Auto-Submit: Ben Wagner <bungeman@google.com>