[graphite] Add a non-threaded PipelineManager
This CL introduces simplified versions of the three main parts of the threaded compilation system:
  GraphicsPipelineHandle - a proxy for a Pipeline/Creation Task
  PipelineCreationTask - a task that can be run to build a Pipeline
  PipelineManager - orchestrates the creation of Pipelines

This CL doesn't change a lot of behavior but just adds overhead.

In DrawPass::prepareResources we create GraphicsPipelineHandles from the GraphicsPipelineDescs and then, immediately, resolve them to GraphicsPipelines.

If the GraphicsPipeline already exists in the cache, it is found and
returned w/in the Handle.

If the GraphicsPipeline doesn't exist then we will create a Task, put it in the TaskMap, blindly execute it in-line, then try to remove it from the TaskMap. This causes a bit of thread-safety danger since, if another Recording is being snapped at the same time or a Precompile is happening (with the same Pipeline), the two threads could share the same task, both race to complete it and then both attempt to remove it from the map. The race to Pipeline completion already currently exists so the main thread-safety risk is in the TaskMap's state.


Bug: b/434712686
Change-Id: Id2a329e598d5e20c561838e53d88e6cd459deda6
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1046996
Reviewed-by: Thomas Smith <thomsmit@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Robert Phillips <robertphillips@google.com>