Mac viewer builds with bazel
Some of the required objective-c files were compiled with ARC but they
do MRC and so didn't compile.

objc_library has a non_arc_srcs list that compiles files with
-fno-objc-arc but using this through the skia_objc_library macro causes
an obscure crash in bazel. So it is reimplemented in the macro using a
nested objc_library that augments its copts.

Change-Id: Ie04f2531ff741ac0d209a6bdca2f793c488e63ab
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1051339
Commit-Queue: Kaylee Lubick <kjlubick@google.com>
Reviewed-by: Kaylee Lubick <kjlubick@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>