Remove drawing element that has undefined behavior
The removed draw was explicitly calling setConvexity (a private call)
on a path that was not convex -- this means as we happen to change
the internal flow, sometimes this draws differently -- but since we've
forced the path into an inconsistent state, we can't really trust any
output. Hence we're removing this.

Change-Id: Icbca45f1c60b8b1ba1ad72fb6d51eea2ec532ba1
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1052596
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Mike Reed <mike@reedtribe.org>
Reviewed-by: Robert Phillips <robertphillips@google.com>