Fold direction into convexity enum
With a single (expanded) enum, we only record direction if we are
also determined to be convex -- i.e. a concave or undetermined
path does not cache any direction information. (note: a convex path
may also have a known degenerate direction... e.g. a line).

Here are the working assumptions

- CPU only cares about first direction in StrokeAndFill, which is
  not critical (chrome may never use it).
- GPU only cares about first direction for convex paths.

The previous impl that stored convexity and direction separately had
odd order dependencies -- computing convexity might alsoset the
direction (or not) -- but computing direction did not try to compute
convexity.

The new impl ensures that they are computed synchronously. Once
convexity has been determined, the direction is now settled: either known or degenerate.

It is expected that with this change, we will be able to remove the
SkPathFirstDirection enum entirely....

Change-Id: Ibabf9d388cfdee7204590a1f318536170b6c64e7
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1049417
Reviewed-by: Kaylee Lubick <kjlubick@google.com>
Commit-Queue: Mike Reed <mike@reedtribe.org>
Reviewed-by: Robert Phillips <robertphillips@google.com>