skhdr::Metadata: Add CLLI and MDCV metadata and parsing
Add the skhdr namespace and add the skhdr::Metadata structure, which
includes all HDR metadata.

Add the structures skhdr::ContentLightLevelInformation and
skhdr::MasteringDisplayColorVolume, with code to parse and print, and
tests for this.

Of note is that this mirrors several gfx::HDRMetadata structures. The
goal is to replace those structures with SkHdrMetadata structures.

Bug: 376550658
Change-Id: I38330824603e12efc0ab37bf3872050acdd41040
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1053237
Reviewed-by: Christopher Cameron <ccameron@google.com>
Commit-Queue: Christopher Cameron <ccameron@google.com>
Reviewed-by: Kaylee Lubick <kjlubick@google.com>