skhdr::Metadata: Add CLLI and MDCV metadata and parsing
Add the skhdr namespace and add the skhdr::Metadata structure, which
includes all HDR metadata.

Add the structures skhdr::ContentLightLevelInformation and
skhdr::MasteringDisplayColorVolume, with code to parse and print, and
tests for this.

Of note is that this mirrors several gfx::HDRMetadata structures. The
goal is to replace those structures with SkHdrMetadata structures.

Bug: 376550658
Change-Id: I9ced6cbdd8d20e7370420f303cd023fc1e050341
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1055616
Reviewed-by: Kaylee Lubick <kjlubick@google.com>
Commit-Queue: Christopher Cameron <ccameron@google.com>
Reviewed-by: Christopher Cameron <ccameron@google.com>