Assume LazyTexture images have successfully been pinned.
Inspired by a patch from caihui yang [1].

In SkImage_AndroidFactories.cpp, `DeferredFromAHardwareBuffer`
creates one of these LazyTexture images via
`DeferredFromTextureGenerator`. If these are passed into
`PinAsTexture`, we can return true because the data in the image
is already in the GPU cache (or can be there quickly). If we
return false, then [2] fails and we do unnecessary work to
unpin the other images (evicting them from the cache) and then
not using the pin shortcut to do work ahead of time.

[1] https://skia-review.googlesource.com/c/skia/+/1054976
[2] http://ac/frameworks/base/libs/hwui/pipeline/skia/SkiaDisplayList.cpp?l=100-105

Change-Id: I58b73e508420bd517ed99fa3a9654cafd27bb580
Bug: b/444322854
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1055656
Commit-Queue: Kaylee Lubick <kjlubick@google.com>
Reviewed-by: John Reck <jreck@google.com>