[rustpng] Implement subsets for SkPngRustCodec
For further feature parity with the libpng codec, implement subsets. We
don't support subsets for animated images for now, as our default png
codec doesn't support animated images. This means that our
implementation only has to worry about one frame, and that the frame's
dimensions will be equal to the image dimensions.

For non-interlaced images, we don't have to do much as applyXform() and
SkSwizzler will adjust it correctly. We just have to note which rows to
copy to dst, using member variables fFirstRow and fLastRow. We still do
have to decode each row before fFirstRow, because of the rust crate api.

For interlaced images, we have to decode into a buffer for the full
image. The rust crate expects the full row and whole dst buffer for
calls to expand_interlaced_row(). So we use fPreblendBuffer to read the
image into, then extract a subset into our final dst with
getSubsetFromFullImage(). We also wait before apply any transforms
before decoding the full image, because applyXForm may truncate data we
need later (I think this is also how SkPngCodec works as well).

Adds a test to SkPngRustDecoderTest.cpp, which tests subsets on an
interlaced and non-interlaced image. This is based on test_in_stripes()
from CodecTest.cpp. This is another step in being able to run all
skia/tests through SkPngRustCodec, but for now I have added this test.

Note: If we want to support subsets for apng in the future, we can add
another field to DecodingState, to have both fInterlacedBuffer and
fPreblendBuffer.

Bug: 362830091
Bug: 356875275
Change-Id: I223b4f867efa3526b02fe3cafe87e00a20cd34f8
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1054479
Reviewed-by: Kaylee Lubick <kjlubick@google.com>
Commit-Queue: Daniel Dilan <danieldilan@google.com>
Reviewed-by: Łukasz Anforowicz <lukasza@google.com>