SkCodec/SkPngCodec/SkPngEncoder: Add HDR metadata support
To SkEncodedInfo, add an skhdr::Metadata item.

To SkCodec, add an accessor for the skhdr::Metadata in its
SkEncodedInfo.

In SkPngCompositeChunkReader, which currently reads the gainmap chunks,
add support to read the MDCV and CLLI chunks. Store an skhdr::Metadata
member in SkPngCompositeChunkReader and propagate it to the
SkEncodedInfo of the SkPngCodec.

In SkPngEncoderMgr, rename setV0Gainmap to setHdrMetadata, which will
write the MDCV and CLLI chunks (in addition to the gainmap chunks).
Slightly clean up this area.

Do not add tests yet. They come with the Rust implementation.

Bug: 376550658
Change-Id: Ibf90e237202b00d589e06fe05d25d59b99256ad0
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1055197
Reviewed-by: Christopher Cameron <ccameron@google.com>
Commit-Queue: ccameron chromium <ccameron@chromium.org>
Owners-Override: Kaylee Lubick <kjlubick@google.com>
Commit-Queue: Christopher Cameron <ccameron@google.com>
Reviewed-by: Kaylee Lubick <kjlubick@google.com>