Migrate Ubuntu18 jobs onto Ubuntu24 machines
Our compiles on Ubuntu18 used to be in docker containers
(with an Ubuntu18 image). These new jobs (matching the
behavior of existing tasks) will be done on the "bare metal"
machine.

This had to update the version of linux_vulkan_sdk because now
that we have drivers which support 1.4, we needed the sdk to also
support 1.4.

Bug: b/445141456
Change-Id: I2b33487a709bf98a677d222f890bd13ea7ae60bf
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1056396
Auto-Submit: Kaylee Lubick <kjlubick@google.com>
Commit-Queue: Nicolette Prevost <nicolettep@google.com>
Reviewed-by: Nicolette Prevost <nicolettep@google.com>