[rustpng] Expand test coverage
Add a test based on this chromium test [1]. Makes sense to have it in
Skia now since we newly support APNG and this is a standard case.
Additionally, add compile time flag in CodecTest for test case which
explicitly uses libpng for testing. If we are running skia/tests through
SkPngRustCodec, we don't want to trigger it.

[1]
https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/platform/image-decoders/png/png_image_decoder_test.cc;drc=25ff1af74a5bac6eea8229c71501d960e65ccb35;l=1939

Bug: 356875275
Change-Id: I45213f51dc2daab6b7be61a18897b72f8558745e
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1057577
Commit-Queue: Daniel Dilan <danieldilan@google.com>
Reviewed-by: Kaylee Lubick <kjlubick@google.com>